fix: P1-T5 - 对齐 /generate 协议 strategy 并增强服务端容错
