+++ piano_duet_server/server/engines/placeholder_inference.py
@@ -4,7 +4,7 @@ import random
 from dataclasses import dataclass
 from typing import Protocol
 
-from ..api.protocol import DialogueNote, GenerateParams
+from ..api.protocol import DialogueNote, GenerateParams, legalize_notes
 
 
 class InferenceEngineProtocol(Protocol):
@@ -14,29 +14,6 @@ class InferenceEngineProtocol(Protocol):
         params: GenerateParams,
         session_id: str | None,
     ) -> list[DialogueNote]: ...
-
-
-def _clamp_int(value: int, lower: int, upper: int) -> int:
-    return max(lower, min(upper, value))
-
-
-def _clamp_float(value: float, lower: float, upper: float) -> float:
-    return max(lower, min(upper, value))
-
-
-def _legalize_notes(notes: list[DialogueNote]) -> list[DialogueNote]:
-    legalized: list[DialogueNote] = []
-    for note in notes:
-        legalized.append(
-            DialogueNote(
-                note=_clamp_int(int(note.note), 0, 127),
-                velocity=_clamp_int(int(note.velocity), 0, 127),
-                time=_clamp_float(float(note.time), 0.0, 1_000_000.0),
-                duration=_clamp_float(float(note.duration), 0.01, 1_000_000.0),
-            )
-        )
-    legalized.sort(key=lambda item: (item.time, item.note))
-    return legalized
 
 
 @dataclass(frozen=True)
@@ -57,8 +34,8 @@ class PlaceholderInferenceEngine:
 
         rng = random.Random(params.seed if params.seed is not None else 0)
         pattern = [0, 4, 7, 12, 7, 4]
-        base = _clamp_int(last_note, 36, 84)
-        velocity = _clamp_int(max(40, min(110, last_velocity)), 0, 127)
+        base = max(36, min(84, last_note))
+        velocity = max(0, min(127, max(40, min(110, last_velocity))))
 
         reply: list[DialogueNote] = []
         step = 0.18
@@ -67,14 +44,14 @@ class PlaceholderInferenceEngine:
             pitch = base + pattern[i] + rng.choice([0, 0, 0, 12])
             reply.append(
                 DialogueNote(
-                    note=_clamp_int(pitch, 0, 127),
+                    note=int(pitch),
                     velocity=velocity,
                     time=float(i) * step,
                     duration=dur,
                 )
             )
 
-        return _legalize_notes(reply)
+        return legalize_notes(reply)
 
 
 _ENGINE: InferenceEngineProtocol | None = None
